requirements.txt
langchain
langchain-openai
chromadb
python-dotenv
tiktoken
langchain_community